# Elude parser allocation