--- a/ext/json/ext/parser/parser.c
+++ b/ext/json/ext/parser/parser.c
@@ -128,7 +128,7 @@ static char *JSON_parse_object(JSON_Parser *json, char *p, char *pe, VALUE *resu
         rb_raise(eNestingError, "nesting of %d is too deep", current_nesting);
     }
 
-    *result = NIL_P(object_class) ? rb_hash_new() : rb_class_new_instance(0, 0, object_class);
+    *result = object_class ?  rb_class_new_instance(0, 0, object_class) :rb_hash_new();
 
 
 #line 135 "parser.c"
@@ -253,11 +253,11 @@ tr11:
         if (np == NULL) {
             p--; {p++; cs = 9; goto _out;}
         } else {
-            if (NIL_P(json->object_class)) {
+            if (json->object_class) {
+                rb_funcall(*result, i_aset, 2, last_name, v);
+            } else {
                 OBJ_FREEZE(last_name);
                 rb_hash_aset(*result, last_name, v);
-            } else {
-                rb_funcall(*result, i_aset, 2, last_name, v);
             }
             {p = (( np))-1;}
         }
@@ -466,10 +466,10 @@ case 26:
     if (cs >= JSON_object_first_final) {
         if (json->create_additions) {
             VALUE klassname;
-            if (NIL_P(json->object_class)) {
-              klassname = rb_hash_aref(*result, json->create_id);
+            if (json->object_class) {
+                klassname = rb_funcall(*result, i_aref, 1, json->create_id);
             } else {
-              klassname = rb_funcall(*result, i_aref, 1, json->create_id);
+                klassname = rb_hash_aref(*result, json->create_id);
             }
             if (!NIL_P(klassname)) {
                 VALUE klass = rb_funcall(mJSON, i_deep_const_get, 1, klassname);
@@ -1141,7 +1141,7 @@ case 7:
     if (cs >= JSON_float_first_final) {
         VALUE mod = Qnil;
         ID method_id = 0;
-        if (!NIL_P(json->decimal_class)) {
+        if (json->decimal_class) {
             if (rb_respond_to(json->decimal_class, i_try_convert)) {
                 mod = json->decimal_class;
                 method_id = i_try_convert;
@@ -1208,7 +1208,7 @@ static char *JSON_parse_array(JSON_Parser *json, char *p, char *pe, VALUE *resul
     if (json->max_nesting && current_nesting > json->max_nesting) {
         rb_raise(eNestingError, "nesting of %d is too deep", current_nesting);
     }
-    *result = NIL_P(array_class) ? rb_ary_new() : rb_class_new_instance(0, 0, array_class);
+    *result = array_class ? rb_class_new_instance(0, 0, array_class) : rb_ary_new();
 
 
 #line 1215 "parser.c"
@@ -1264,10 +1264,10 @@ tr2:
         if (np == NULL) {
             p--; {p++; cs = 3; goto _out;}
         } else {
-            if (NIL_P(json->array_class)) {
-                rb_ary_push(*result, v);
-            } else {
+            if (json->array_class) {
                 rb_funcall(*result, i_leftshift, 1, v);
+            } else {
+                rb_ary_push(*result, v);
             }
             {p = (( np))-1;}
         }
@@ -1788,6 +1788,103 @@ static VALUE convert_encoding(VALUE source)
   return rb_funcall(source, i_encode, 1, Encoding_UTF_8);
 }
 
+static void parser_init(JSON_Parser *json, VALUE source, VALUE opts)
+{
+    if (json->Vsource) {
+        rb_raise(rb_eTypeError, "already initialized instance");
+    }
+
+    json->fbuffer.initial_length = FBUFFER_INITIAL_LENGTH_DEFAULT;
+    json->max_nesting = 100;
+
+    if (!NIL_P(opts)) {
+        Check_Type(opts, T_HASH);
+        if (RHASH_SIZE(opts) > 0) {
+            VALUE tmp = ID2SYM(i_max_nesting);
+            if (option_given_p(opts, tmp)) {
+                VALUE max_nesting = rb_hash_aref(opts, tmp);
+                if (RTEST(max_nesting)) {
+                    Check_Type(max_nesting, T_FIXNUM);
+                    json->max_nesting = FIX2INT(max_nesting);
+                } else {
+                    json->max_nesting = 0;
+                }
+            } else {
+                json->max_nesting = 100;
+            }
+            tmp = ID2SYM(i_allow_nan);
+            if (option_given_p(opts, tmp)) {
+                json->allow_nan = RTEST(rb_hash_aref(opts, tmp)) ? 1 : 0;
+            } else {
+                json->allow_nan = 0;
+            }
+            tmp = ID2SYM(i_symbolize_names);
+            if (option_given_p(opts, tmp)) {
+                json->symbolize_names = RTEST(rb_hash_aref(opts, tmp)) ? 1 : 0;
+            } else {
+                json->symbolize_names = 0;
+            }
+            tmp = ID2SYM(i_freeze);
+            if (option_given_p(opts, tmp)) {
+                json->freeze = RTEST(rb_hash_aref(opts, tmp)) ? 1 : 0;
+            } else {
+                json->freeze = 0;
+            }
+            tmp = ID2SYM(i_create_additions);
+            if (option_given_p(opts, tmp)) {
+                tmp = rb_hash_aref(opts, tmp);
+                if (NIL_P(tmp)) {
+                    json->create_additions = 1;
+                    json->deprecated_create_additions = 1;
+                } else {
+                    json->create_additions = RTEST(tmp);
+                    json->deprecated_create_additions = 0;
+                }
+            }
+
+            if (json->symbolize_names && json->create_additions) {
+                rb_raise(rb_eArgError,
+                    "options :symbolize_names and :create_additions cannot be "
+                    " used in conjunction");
+            }
+            tmp = ID2SYM(i_create_id);
+            if (option_given_p(opts, tmp)) {
+                json->create_id = rb_hash_aref(opts, tmp);
+            } else {
+                json->create_id = rb_funcall(mJSON, i_create_id, 0);
+            }
+            tmp = ID2SYM(i_object_class);
+            if (option_given_p(opts, tmp)) {
+                json->object_class = rb_hash_aref(opts, tmp);
+                if (NIL_P(json->object_class)) json->object_class = Qfalse;
+            }
+            tmp = ID2SYM(i_array_class);
+            if (option_given_p(opts, tmp)) {
+                json->array_class = rb_hash_aref(opts, tmp);
+                if (NIL_P(json->array_class)) json->array_class = Qfalse;
+            }
+
+            tmp = ID2SYM(i_decimal_class);
+            if (option_given_p(opts, tmp)) {
+                json->decimal_class = rb_hash_aref(opts, tmp);
+                if (NIL_P(json->decimal_class)) json->decimal_class = Qfalse;
+            }
+
+            tmp = ID2SYM(i_match_string);
+            if (option_given_p(opts, tmp)) {
+                VALUE match_string = rb_hash_aref(opts, tmp);
+                json->match_string = RTEST(match_string) ? match_string : Qfalse;
+            }
+        }
+    }
+
+    source = convert_encoding(StringValue(source));
+    StringValue(source);
+    json->len = RSTRING_LEN(source);
+    json->source = RSTRING_PTR(source);
+    json->Vsource = source;
+}
+
 /*
  * call-seq: new(source, opts => {})
  *
@@ -1822,122 +1919,16 @@ static VALUE convert_encoding(VALUE source)
  */
 static VALUE cParser_initialize(int argc, VALUE *argv, VALUE self)
 {
-    VALUE source, opts;
     GET_PARSER_INIT;
 
-    if (json->Vsource) {
-        rb_raise(rb_eTypeError, "already initialized instance");
-    }
-
     rb_check_arity(argc, 1, 2);
-    source = argv[0];
-    opts = Qnil;
-    if (argc == 2) {
-        opts = argv[1];
-        Check_Type(argv[1], T_HASH);
-        if (RHASH_SIZE(argv[1]) > 0) {
-            opts = argv[1];
-        }
-    }
 
-    if (!NIL_P(opts)) {
-        VALUE tmp = ID2SYM(i_max_nesting);
-        if (option_given_p(opts, tmp)) {
-            VALUE max_nesting = rb_hash_aref(opts, tmp);
-            if (RTEST(max_nesting)) {
-                Check_Type(max_nesting, T_FIXNUM);
-                json->max_nesting = FIX2INT(max_nesting);
-            } else {
-                json->max_nesting = 0;
-            }
-        } else {
-            json->max_nesting = 100;
-        }
-        tmp = ID2SYM(i_allow_nan);
-        if (option_given_p(opts, tmp)) {
-            json->allow_nan = RTEST(rb_hash_aref(opts, tmp)) ? 1 : 0;
-        } else {
-            json->allow_nan = 0;
-        }
-        tmp = ID2SYM(i_symbolize_names);
-        if (option_given_p(opts, tmp)) {
-            json->symbolize_names = RTEST(rb_hash_aref(opts, tmp)) ? 1 : 0;
-        } else {
-            json->symbolize_names = 0;
-        }
-        tmp = ID2SYM(i_freeze);
-        if (option_given_p(opts, tmp)) {
-            json->freeze = RTEST(rb_hash_aref(opts, tmp)) ? 1 : 0;
-        } else {
-            json->freeze = 0;
-        }
-        tmp = ID2SYM(i_create_additions);
-        if (option_given_p(opts, tmp)) {
-            tmp = rb_hash_aref(opts, tmp);
-            if (NIL_P(tmp)) {
-                json->create_additions = 1;
-                json->deprecated_create_additions = 1;
-            } else {
-                json->create_additions = RTEST(tmp);
-                json->deprecated_create_additions = 0;
-            }
-        }
-
-        if (json->symbolize_names && json->create_additions) {
-            rb_raise(rb_eArgError,
-                "options :symbolize_names and :create_additions cannot be "
-                " used in conjunction");
-        }
-        tmp = ID2SYM(i_create_id);
-        if (option_given_p(opts, tmp)) {
-            json->create_id = rb_hash_aref(opts, tmp);
-        } else {
-            json->create_id = rb_funcall(mJSON, i_create_id, 0);
-        }
-        tmp = ID2SYM(i_object_class);
-        if (option_given_p(opts, tmp)) {
-            json->object_class = rb_hash_aref(opts, tmp);
-        } else {
-            json->object_class = Qnil;
-        }
-        tmp = ID2SYM(i_array_class);
-        if (option_given_p(opts, tmp)) {
-            json->array_class = rb_hash_aref(opts, tmp);
-        } else {
-            json->array_class = Qnil;
-        }
-        tmp = ID2SYM(i_decimal_class);
-        if (option_given_p(opts, tmp)) {
-            json->decimal_class = rb_hash_aref(opts, tmp);
-        } else {
-            json->decimal_class = Qnil;
-        }
-        tmp = ID2SYM(i_match_string);
-        if (option_given_p(opts, tmp)) {
-            VALUE match_string = rb_hash_aref(opts, tmp);
-            json->match_string = RTEST(match_string) ? match_string : Qnil;
-        } else {
-            json->match_string = Qnil;
-        }
-    } else {
-        json->max_nesting = 100;
-        json->allow_nan = 0;
-        json->create_additions = 0;
-        json->create_id = Qnil;
-        json->object_class = Qnil;
-        json->array_class = Qnil;
-        json->decimal_class = Qnil;
-    }
-    source = convert_encoding(StringValue(source));
-    StringValue(source);
-    json->len = RSTRING_LEN(source);
-    json->source = RSTRING_PTR(source);
-    json->Vsource = source;
+    parser_init(json, argv[0], argc == 2 ? argv[1] : Qnil);
     return self;
 }
 
 
-#line 1941 "parser.c"
+#line 1932 "parser.c"
 enum {JSON_start = 1};
 enum {JSON_first_final = 10};
 enum {JSON_error = 0};
@@ -1945,7 +1936,7 @@ enum {JSON_error = 0};
 enum {JSON_en_main = 1};
 
 
-#line 849 "parser.rl"
+#line 840 "parser.rl"
 
 
 /*
@@ -1963,16 +1954,16 @@ static VALUE cParser_parse(VALUE self)
     GET_PARSER;
 
 
-#line 1967 "parser.c"
+#line 1958 "parser.c"
 	{
 	cs = JSON_start;
 	}
 
-#line 866 "parser.rl"
+#line 857 "parser.rl"
     p = json->source;
     pe = p + json->len;
 
-#line 1976 "parser.c"
+#line 1967 "parser.c"
 	{
 	if ( p == pe )
 		goto _test_eof;
@@ -2006,7 +1997,7 @@ st0:
 cs = 0;
 	goto _out;
 tr2:
-#line 841 "parser.rl"
+#line 832 "parser.rl"
 	{
         char *np = JSON_parse_value(json, p, pe, &result, 0);
         if (np == NULL) { p--; {p++; cs = 10; goto _out;} } else {p = (( np))-1;}
@@ -2016,7 +2007,7 @@ st10:
 	if ( ++p == pe )
 		goto _test_eof10;
 case 10:
-#line 2020 "parser.c"
+#line 2011 "parser.c"
 	switch( (*p) ) {
 		case 13: goto st10;
 		case 32: goto st10;
@@ -2105,7 +2096,170 @@ case 9:
 	_out: {}
 	}
 
-#line 869 "parser.rl"
+#line 860 "parser.rl"
+
+    if (cs >= JSON_first_final && p == pe) {
+        return result;
+    } else {
+        raise_parse_error("unexpected token at '%s'", p);
+        return Qnil;
+    }
+}
+
+static VALUE cParser_m_parse(VALUE klass, VALUE source, VALUE opts)
+{
+    char *p, *pe;
+    int cs = EVIL;
+    VALUE result = Qnil;
+
+    JSON_Parser parser = {0};
+    JSON_Parser *json = &parser;
+    parser_init(json, source, opts);
+
+
+#line 2121 "parser.c"
+	{
+	cs = JSON_start;
+	}
+
+#line 880 "parser.rl"
+    p = json->source;
+    pe = p + json->len;
+
+#line 2130 "parser.c"
+	{
+	if ( p == pe )
+		goto _test_eof;
+	switch ( cs )
+	{
+st1:
+	if ( ++p == pe )
+		goto _test_eof1;
+case 1:
+	switch( (*p) ) {
+		case 13: goto st1;
+		case 32: goto st1;
+		case 34: goto tr2;
+		case 45: goto tr2;
+		case 47: goto st6;
+		case 73: goto tr2;
+		case 78: goto tr2;
+		case 91: goto tr2;
+		case 102: goto tr2;
+		case 110: goto tr2;
+		case 116: goto tr2;
+		case 123: goto tr2;
+	}
+	if ( (*p) > 10 ) {
+		if ( 48 <= (*p) && (*p) <= 57 )
+			goto tr2;
+	} else if ( (*p) >= 9 )
+		goto st1;
+	goto st0;
+st0:
+cs = 0;
+	goto _out;
+tr2:
+#line 832 "parser.rl"
+	{
+        char *np = JSON_parse_value(json, p, pe, &result, 0);
+        if (np == NULL) { p--; {p++; cs = 10; goto _out;} } else {p = (( np))-1;}
+    }
+	goto st10;
+st10:
+	if ( ++p == pe )
+		goto _test_eof10;
+case 10:
+#line 2174 "parser.c"
+	switch( (*p) ) {
+		case 13: goto st10;
+		case 32: goto st10;
+		case 47: goto st2;
+	}
+	if ( 9 <= (*p) && (*p) <= 10 )
+		goto st10;
+	goto st0;
+st2:
+	if ( ++p == pe )
+		goto _test_eof2;
+case 2:
+	switch( (*p) ) {
+		case 42: goto st3;
+		case 47: goto st5;
+	}
+	goto st0;
+st3:
+	if ( ++p == pe )
+		goto _test_eof3;
+case 3:
+	if ( (*p) == 42 )
+		goto st4;
+	goto st3;
+st4:
+	if ( ++p == pe )
+		goto _test_eof4;
+case 4:
+	switch( (*p) ) {
+		case 42: goto st4;
+		case 47: goto st10;
+	}
+	goto st3;
+st5:
+	if ( ++p == pe )
+		goto _test_eof5;
+case 5:
+	if ( (*p) == 10 )
+		goto st10;
+	goto st5;
+st6:
+	if ( ++p == pe )
+		goto _test_eof6;
+case 6:
+	switch( (*p) ) {
+		case 42: goto st7;
+		case 47: goto st9;
+	}
+	goto st0;
+st7:
+	if ( ++p == pe )
+		goto _test_eof7;
+case 7:
+	if ( (*p) == 42 )
+		goto st8;
+	goto st7;
+st8:
+	if ( ++p == pe )
+		goto _test_eof8;
+case 8:
+	switch( (*p) ) {
+		case 42: goto st8;
+		case 47: goto st1;
+	}
+	goto st7;
+st9:
+	if ( ++p == pe )
+		goto _test_eof9;
+case 9:
+	if ( (*p) == 10 )
+		goto st1;
+	goto st9;
+	}
+	_test_eof1: cs = 1; goto _test_eof;
+	_test_eof10: cs = 10; goto _test_eof;
+	_test_eof2: cs = 2; goto _test_eof;
+	_test_eof3: cs = 3; goto _test_eof;
+	_test_eof4: cs = 4; goto _test_eof;
+	_test_eof5: cs = 5; goto _test_eof;
+	_test_eof6: cs = 6; goto _test_eof;
+	_test_eof7: cs = 7; goto _test_eof;
+	_test_eof8: cs = 8; goto _test_eof;
+	_test_eof9: cs = 9; goto _test_eof;
+
+	_test_eof: {}
+	_out: {}
+	}
+
+#line 883 "parser.rl"
 
     if (cs >= JSON_first_final && p == pe) {
         return result;
@@ -2183,6 +2337,8 @@ void Init_parser(void)
     rb_define_method(cParser, "initialize", cParser_initialize, -1);
     rb_define_method(cParser, "parse", cParser_parse, 0);
     rb_define_method(cParser, "source", cParser_source, 0);
+
+    rb_define_singleton_method(cParser, "parse", cParser_m_parse, 2);
 
     CNaN = rb_const_get(mJSON, rb_intern("NaN"));
     rb_gc_register_mark_object(CNaN);