--- a/lib/json/common.rb
+++ b/lib/json/common.rb
@@ -207,16 +207,7 @@ module JSON
   #   JSON.parse('')
   #
   def parse(source, opts = nil)
-    if opts.nil?
-      Parser.new(source).parse
-    else
-      # NB: The ** shouldn't be required, but we have to deal with
-      # different versions of the `json` and `json_pure` gems being
-      # loaded concurrently.
-      # Prior to 2.7.3, `JSON::Ext::Parser` would only take kwargs.
-      # Ref: https://github.com/ruby/json/issues/650
-      Parser.new(source, **opts).parse
-    end
+    Parser.parse(source, opts)
   end
 
   # :call-seq: