--- a/lib/json/pure/parser.rb
+++ b/lib/json/pure/parser.rb
@@ -48,6 +48,21 @@ module JSON
 
       UNPARSED = Object.new.freeze
 
+      class << self
+        def parse(source, opts = nil)
+          if opts.nil?
+            new(source).parse
+          else
+            # NB: The ** shouldn't be required, but we have to deal with
+            # different versions of the `json` and `json_pure` gems being
+            # loaded concurrently.
+            # Prior to 2.7.3, `JSON::Ext::Parser` would only take kwargs.
+            # Ref: https://github.com/ruby/json/issues/650
+            new(source, **opts).parse
+          end
+        end
+      end
+
       # Creates a new JSON::Pure::Parser instance for the string _source_.
       #
       # It will be configured by the _opts_ hash. _opts_ can have the following