Similar to the optimizations done for the generator, but here it's even simpler because we never need to spill the `Parser` object on the heap.

We also start the parsing buffer with 512B of stack memory, so that in most case no allocations other than the returned value are necessary.

This mostly help on micro-benchmarks, but isn't detrimental to more real world benchmarks.

Before:


```
== Parsing small hash (65 bytes)
ruby 3.3.4 (2024-07-09 revision be1089c8ec) +YJIT [arm64-darwin23]
Warming up --------------------------------------
                json   188.233k i/100ms
                  oj   213.985k i/100ms
           oj strict   242.564k i/100ms
          Oj::Parser   448.682k i/100ms
           rapidjson   291.925k i/100ms
Calculating -------------------------------------
                json      1.983M (± 0.5%) i/s  (504.32 ns/i) -      9.976M in   5.031352s
                  oj      2.334M (± 0.2%) i/s  (428.48 ns/i) -     11.769M in   5.042839s
           oj strict      2.689M (± 0.2%) i/s  (371.85 ns/i) -     13.584M in   5.051044s
          Oj::Parser      4.662M (± 1.2%) i/s  (214.50 ns/i) -     23.331M in   5.005414s
           rapidjson      3.110M (± 0.7%) i/s  (321.57 ns/i) -     15.764M in   5.069531s

Comparison:
                json:  1982878.1 i/s
          Oj::Parser:  4661924.8 i/s - 2.35x  faster
           rapidjson:  3109722.2 i/s - 1.57x  faster
           oj strict:  2689277.0 i/s - 1.36x  faster
                  oj:  2333852.9 i/s - 1.18x  faster
```

After:

```
== Parsing small hash (65 bytes)
ruby 3.3.4 (2024-07-09 revision be1089c8ec) [arm64-darwin23]
Warming up --------------------------------------
                json   304.417k i/100ms
                  oj   219.431k i/100ms
           oj strict   254.532k i/100ms
          Oj::Parser   431.309k i/100ms
           rapidjson   281.703k i/100ms
Calculating -------------------------------------
                json      3.046M (± 0.1%) i/s  (328.25 ns/i) -     15.525M in   5.096243s
                  oj      2.225M (± 0.2%) i/s  (449.50 ns/i) -     11.191M in   5.030429s
           oj strict      2.553M (± 0.5%) i/s  (391.75 ns/i) -     12.981M in   5.085538s
          Oj::Parser      4.280M (± 0.8%) i/s  (233.64 ns/i) -     21.565M in   5.038834s
           rapidjson      2.826M (± 0.3%) i/s  (353.83 ns/i) -     14.367M in   5.083480s

Comparison:
                json:  3046420.8 i/s
          Oj::Parser:  4280132.7 i/s - 1.40x  faster
           rapidjson:  2826209.4 i/s - 1.08x  slower
           oj strict:  2552619.7 i/s - 1.19x  slower
                  oj:  2224670.7 i/s - 1.37x  slower
```