@byroot 2024-11-04
You mean the source? It's right there in the repo: https://github.com/ruby/json/tree/master/benchmark